backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain
weasyprint
python-multipart
openai
langchain-openai
python-dotenv
pypdf
